+++ brest/multimeters/__init__.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+    brest.multimeters.__init__
+    ~~~~~~~~~~~~~~~~~~~~~~~
+
+    This file signifies that 'multimeters' is a module.
+
+    :copyright: 2024 Bender Robotics
+"""
+
+from .multimeters import Multimeters
+from .multicomp import Multicomp
+
+__all__ = ['Multimeters', 'Multicomp']
+
+__version__ = '0.0.18'